# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface UserProfile {
  id: number;
  clerkUserId: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  displayName?: string | null;
  createdAt: string;
  lastSeenAt: string;
}

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  displayName?: string;
}

export interface AccountStats {
  memberSince: string;
  lastSeenAt: string;
  signInCount: number;
  activityCount: number;
}

export interface ActivityEvent {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}