# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: Current user profile operations
  - name: activity
    description: Account activity operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /me:
    get:
      operationId: getMe
      tags: [profile]
      summary: Get current user profile
      description: Returns the authenticated user's profile, provisioning it on first access.
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMe
      tags: [profile]
      summary: Update current user profile
      description: Updates the authenticated user's editable profile fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/stats:
    get:
      operationId: getMyStats
      tags: [profile]
      summary: Get current user account summary
      description: Returns aggregate account stats for the authenticated user.
      responses:
        "200":
          description: Account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /activity:
    get:
      operationId: getActivity
      tags: [activity]
      summary: List recent account activity
      description: Returns the authenticated user's recent account activity events, newest first.
      responses:
        "200":
          description: Recent activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        email:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
      required:
        - id
        - clerkUserId
        - createdAt
        - lastSeenAt
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 80
    AccountStats:
      type: object
      properties:
        memberSince:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        signInCount:
          type: integer
        activityCount:
          type: integer
      required:
        - memberSince
        - lastSeenAt
        - signInCount
        - activityCount
    ActivityEvent:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - createdAt