# lib/api-zod/src/generated/types/accountStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountStats {
  memberSince: Date;
  lastSeenAt: Date;
  signInCount: number;
  activityCount: number;
}