# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  clerkUserId: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  displayName?: string | null;
  createdAt: Date;
  lastSeenAt: Date;
}